% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLPM.R
\name{DMLPM}
\alias{DMLPM}
\title{Implements the DMLPM model}
\usage{
DMLPM(
  num_samples = 2000,
  burn = 10000,
  thin = 10,
  post_pred_thin_factor = 5,
  num_chains = 1,
  Y,
  sigma2_0 = sigma2_0,
  sigma2_a = 1,
  sigma2_z = 1,
  alpha_0 = 1,
  Z_0 = NULL,
  eta_alpha = 1,
  eta_z = 1,
  tune = TRUE,
  delta_alpha = 0.1,
  delta_z = 0.1,
  AR_min = 0.2,
  AR_max = 0.4,
  discard = FALSE
)
}
\arguments{
\item{discard}{}
}
\description{
Implements the DMLPM model
}
