% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise_alpha}
\alias{initialise_alpha}
\title{Initialise the alphas using the regression method}
\usage{
initialise_alpha(Y, Z)
}
\arguments{
\item{Z}{}
}
\description{
Initialise the alphas using the regression method
}
