% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{log_posterior}
\alias{log_posterior}
\title{Compute the full log-posterior}
\usage{
log_posterior(
  Y,
  Z,
  alpha,
  sigma2_0 = sigma2_0,
  sigma2_a = sigma2_a,
  sigma2_z = sigma2_z
)
}
\arguments{
\item{sigma2_z}{}
}
\description{
Compute the full log-posterior
}
