% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{log_posterior_alpha_kt}
\alias{log_posterior_alpha_kt}
\title{Compute the contribution to the log-posterior after updating at a given time t and layer k}
\usage{
log_posterior_alpha_kt(
  Y_tk,
  D_t,
  alpha,
  sigma2_0 = sigma2_0,
  sigma2_a = sigma2_a,
  sigma2_z = sigma2_z,
  k = 1,
  t = 1
)
}
\arguments{
\item{t}{}
}
\description{
Compute the contribution to the log-posterior after updating at a given time t and layer k
}
