% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{LPM_gen_poisson}
\alias{LPM_gen_poisson}
\title{Simulate networks from Poisson LPM given latent positions and intercept}
\usage{
LPM_gen_poisson(Z, alpha)
}
\arguments{
\item{Z}{Matrix of latent positions}

\item{alpha}{Intercept parameter}
}
\description{
Simulate networks from Poisson LPM given latent positions and intercept
}
