% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{netpois}
\alias{netpois}
\title{Used for initialising alpha, see Gwee's paper and code where he uses this idea}
\usage{
netpois(y, x, intercept = TRUE, mode = "digraph", diag = FALSE)
}
\arguments{
\item{diag}{}
}
\description{
Used for initialising alpha, see Gwee's paper and code where he uses this idea
}
